+++ go.mod
@@ -1,3 +1,5 @@
 module endocode
 
 go 1.18
+
+require github.com/go-chi/chi/v5 v5.0.7
