--- server.go
@@ -1,18 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"net/http"
-)
-
-func main() {
-	fmt.Printf("Starting server at port 8080\n")
-	if err := http.ListenAndServe(":8080", nil); err != nil {
-		log.Fatal(err)
-
-		http.HandleFunc("/helloworld", func(w http.ResponseWriter, r *http.Request) {
-			fmt.Fprintf(w, "Hello Stranger")
-		})
-	}
-}
